requirements.txt
gradio
bittensor
matplotlib
pandas
datasets
tqdm
redis
python-dotenv